Changes to Scene, removed ModelResourceManager
Scene now encapsulates the environment. When a new level is to be loaded,
simply call the Create3DLevel()-method passing in the new level as an
argument. This will delete the old environment and create a new one. The
camera is moved to the Pacman Spawn Position loaded from the level.
TODO: Scene shall also contain all pellets, ghosts etc and draw them.

Removed the superfluous ModelResourceManager, as
D3DResourceManager<Resources::StaticModelData> already does exactly the
same thing. Old code is left but commeted out for now (in case someone
does not agree with the changes)